+++ example/integration_test/plugin_integration_test.dart
@@ -12,7 +12,6 @@ import 'package:permission_handler/permission_handler.dart';
 const notificationDetails = NotificationDetails(
   channelId: 'foreground_service_isolate',
   channelName: 'Foreground Service Isolate',
-  channelDescription: 'Foreground Service Isolate',
   id: 1,
   contentTitle: 'Foreground Service Isolate',
   contentText: 'Running...',